Fix sign comparison warning in mg_get_ports
ctx->num_listening_sockets is a signed int. i is a size_t (unsigned long).

+++ src/civetweb.c
@@ -710,7 +710,7 @@ const char *mg_get_option(const struct mg_context *ctx, const char *name)
 void mg_get_ports(const struct mg_context *ctx, size_t size, int* ports, int* ssl)
 {
     size_t i;
-    for (i = 0; i < size && i < ctx->num_listening_sockets; i++)
+    for (i = 0; i < size && i < (size_t)ctx->num_listening_sockets; i++)
     {
         ssl[i] = ctx->listening_sockets[i].is_ssl;
         ports[i] = ctx->listening_ports[i];
